chore(dashboard): remove obsolete js widget
