+++ lib/AppInfo/Application.php
@@ -21,9 +21,7 @@ use OCA\Mail\Contracts\IMailTransmission;
 use OCA\Mail\Contracts\ITrustedSenderService;
 use OCA\Mail\Contracts\IUserPreferences;
 use OCA\Mail\Dashboard\ImportantMailWidget;
-use OCA\Mail\Dashboard\ImportantMailWidgetV2;
 use OCA\Mail\Dashboard\UnreadMailWidget;
-use OCA\Mail\Dashboard\UnreadMailWidgetV2;
 use OCA\Mail\Events\BeforeImapClientCreated;
 use OCA\Mail\Events\DraftMessageCreatedEvent;
 use OCA\Mail\Events\DraftSavedEvent;
@@ -70,7 +68,6 @@ use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
-use OCP\Dashboard\IAPIWidgetV2;
 use OCP\DB\Events\AddMissingIndicesEvent;
 use OCP\IServerContainer;
 use OCP\Search\IFilteringProvider;
@@ -150,13 +147,8 @@ class Application extends App implements IBootstrap {
 		$context->registerMiddleWare(ErrorMiddleware::class);
 		$context->registerMiddleWare(ProvisioningMiddleware::class);
 
-		if (interface_exists(IAPIWidgetV2::class)) {
-			$context->registerDashboardWidget(ImportantMailWidgetV2::class);
-			$context->registerDashboardWidget(UnreadMailWidgetV2::class);
-		} else {
-			$context->registerDashboardWidget(ImportantMailWidget::class);
-			$context->registerDashboardWidget(UnreadMailWidget::class);
-		}
+		$context->registerDashboardWidget(ImportantMailWidget::class);
+		$context->registerDashboardWidget(UnreadMailWidget::class);
 
 		if (interface_exists(IFilteringProvider::class)) {
 			$context->registerSearchProvider(FilteringProvider::class);
